requirements.txt
# MongoDB Connectors (Stable for Django 3.2)
djongo==1.3.6
pymongo[srv]==3.12.1
dnspython==2.4.2

# Core Django
Django==3.2.25
asgiref==3.8.1
sqlparse==0.5.3
python-decouple==3.8
whitenoise==6.6.0

# Data Science
numpy==1.26.4
numpy-financial==1.0.0
pandas==2.2.2
openpyxl==3.1.2

# App Features & Auth
django-import-export==4.1.1
django-allauth==0.54.0
djangorestframework==3.14.0
django-cors-headers==4.5.0
PyJWT==2.8.0
cryptography==42.0.8
django-widget-tweaks==1.5.0
django-filter==23.5
django-contrib-comments==2.2.0
requests==2.32.3
python-dateutil==2.9.0

# Storage
django-cloudinary-storage==0.3.0
cloudinary==1.41.0